+++ src/android/LocalNotification.java
@@ -120,6 +120,15 @@ public class LocalNotification extends CordovaPlugin {
             return true;
         }
 
+        if (action.equalsIgnoreCase("hasPermission")) {
+            hasPermission(callbackContext);
+            return true;
+        }
+
+        if (action.equalsIgnoreCase("promptForPermission")) {
+            return true;
+        }
+
         if (action.equalsIgnoreCase("deviceready")) {
             cordova.getThreadPool().execute( new Runnable() {
                 public void run() {
@@ -263,6 +272,25 @@ public class LocalNotification extends CordovaPlugin {
         JSONArray pendingIds       = new JSONArray(alarmIds);
 
         callbackContext.success(pendingIds);
+    }
+
+    /**
+     * Informs if the app has the permission to show notifications.
+     *
+     * @param callback
+     *      The function to be exec as the callback
+     */
+    private void hasPermission (final CallbackContext callback) {
+        cordova.getThreadPool().execute(new Runnable() {
+            @Override
+            public void run() {
+                PluginResult result;
+
+                result = new PluginResult(PluginResult.Status.OK, true);
+
+                callback.sendPluginResult(result);
+            }
+        });
     }
 
     /**