+++ www/local-notification.js
@@ -169,6 +169,29 @@ LocalNotification.prototype = {
     },
 
     /**
+     * Informs if the app has the permission to show badges.
+     *
+     * @param {Function} callback
+     *      The function to be exec as the callback
+     * @param {Object?} scope
+     *      The callback function's scope
+     */
+    hasPermission: function (callback, scope) {
+        var fn = function (badge) {
+            callback.call(scope || this, badge);
+        };
+
+        cordova.exec(fn, null, 'LocalNotification', 'hasPermission', []);
+    },
+
+    /**
+     * Ask for permission to show badges if not already granted.
+     */
+    promptForPermission: function () {
+        cordova.exec(null, null, 'LocalNotification', 'promptForPermission', []);
+    },
+
+    /**
      * Occurs when a notification was added.
      *
      * @param {String} id    The ID of the notification