+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## ChangeLog
+
+#### Version 0.7.5 (29.10.2014)
+- [enhancement:] __iOS8 Support__
+- [feature:] New method `hasPermission` to ask if the user has granted to display local notifications.
+- [feature:] New method `promptForPermission` to promt the user to grant permission to display local notifications.
+
 #### Version 0.7.4 (22.03.2014)
 - [bugfix:] Platform specific properties were ignored.
 - [bugfix:] `cancel` may throw an error if the OS returns NIL values (iOS).